Restructured code directories.
