+++ sprites/pellet_data.json
@@ -1,0 +1,36 @@
+{
+	"frames": [
+		{
+			"filename": "pellet_0.png",
+			"frame": {
+				"x": 1,
+				"y": 1,
+				"w": 31,
+				"h": 27
+			},
+			"rotated": false,
+			"trimmed": false,
+			"spriteSourceSize": {
+				"x": 0,
+				"y": 0,
+				"w": 31,
+				"h": 27
+			},
+			"sourceSize": {
+				"w": 31,
+				"h": 27
+			}
+		}
+	],
+	"meta": {
+		"app": "http://www.codeandweb.com/texturepacker",
+		"version": "1.0",
+		"image": "spritesheet.png",
+		"format": "RGBA8888",
+		"size": {
+			"w": 33,
+			"h": 29
+		},
+		"scale": "1"
+	}
+}
